+++ scaler/io/cpp/src/main.hpp
@@ -1,8 +1,8 @@
 #include "common.hpp"
 #include "session.hpp"
-#include "client.hpp"
+#include "network_connector.hpp"
 #include "intra_process.hpp"
 
 #include "session.cpp"
-#include "client.cpp"
+#include "network_connector.cpp"
 #include "intra_process.cpp"
